# Speed up TargetEncoder with vectorization.