Roughly 300X faster and uses less memory.
`fit()` computes category smoothed means with vector ops, stores mapping as (dict of) Series.
`transform()` then just calls Series.map(mapping) to do vectorized lookup with the mapping Series.

Also fixes IndexError when passing DataFrames with non-contiguous indexes.

Also fixes not being able to re-use TargetEncoder by always setting `mapping=None` in `fit()` (you don't want fit to remember anything from previous calls).

```
rows, cats = 1000000, 1000
X = pd.DataFrame({'x': np.random.randint(0, cats, rows).astype(str)})
y = pd.Series(np.random.rand(rows))

# old
%time TargetEncoder().fit_transform(X, y)
CPU times: user 1min 24s, sys: 71.2 ms, total: 1min 24s
Wall time: 1min 24s

# new
%time TargetEncoder().fit_transform(X, y)
CPU times: user 235 ms, sys: 43.6 ms, total: 279 ms
Wall time: 276 ms
```