@janmotl 2018-10-21
Hi @jkleint, couldn't you look at LeaveOneOutEncoder and possibly vectorize it as well? I failed in the attempt.